Changed s3 objects delete condition

+++ webapp/librarymanagementsystem/src/main/java/com/csye6225/lms/controller/DatabaseController.java
@@ -4,6 +4,7 @@ import com.csye6225.lms.service.AmazonS3ImageService;
 import com.csye6225.lms.service.BookService;
 import com.csye6225.lms.service.CustomUserDetailsService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.env.Environment;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -21,12 +22,18 @@ public class DatabaseController {
     @Autowired
     private AmazonS3ImageService s3ImageService;
 
+    @Autowired
+    private Environment environment;
+
+
     @PostMapping(value = "/")
     public ResponseEntity<Object> resetPassword(){
         System.out.println("Deleteing all data");
         bookService.deleteAll();
         userService.deleteAll();
-        s3ImageService.deleteAll();
+        if(environment.getActiveProfiles()[0].equalsIgnoreCase("prod")){
+            s3ImageService.deleteAll();
+        }
         System.out.println("Deleted all data");
         return ResponseEntity.noContent().build();
     }
